refactor(clawdstrike): extract check_api entrypoints from engine.rs
Move the public check_* entrypoints (check_file_access, check_file_write,
check_egress, check_shell, check_mcp_tool, check_untrusted_text,
check_patch, check_action, check_action_report) into check_api.rs as an
impl HushEngine block.

The orchestration helpers they call across the (future) evaluate split
(prepare_origin_context, check_action_report_prepared) and the shared
PreparedContext/PreparedEvaluation types + metadata helpers
(single_result_report, build_report_metadata, report_metadata_for_context)
are widened to pub(crate) for cross-module access. Public API unchanged.

Pure relocation: 825 lib tests + formal-diff-tests pass.
