+++ crates/libs/clawdstrike/src/engine/check_api.rs
@@ -1,0 +1,117 @@
+//! Public `check_*` entrypoints for [`HushEngine`].
+
+use crate::error::{Error, Result};
+use crate::guards::{GuardAction, GuardContext, GuardResult};
+
+use super::types::GuardReport;
+use super::{HushEngine, PreparedEvaluation};
+
+impl HushEngine {
+    /// Check a file access action
+    pub async fn check_file_access(
+        &self,
+        path: &str,
+        context: &GuardContext,
+    ) -> Result<GuardResult> {
+        self.check_action(&GuardAction::FileAccess(path), context)
+            .await
+    }
+
+    /// Check a file write action
+    pub async fn check_file_write(
+        &self,
+        path: &str,
+        content: &[u8],
+        context: &GuardContext,
+    ) -> Result<GuardResult> {
+        self.check_action(&GuardAction::FileWrite(path, content), context)
+            .await
+    }
+
+    /// Check a network egress action
+    pub async fn check_egress(
+        &self,
+        host: &str,
+        port: u16,
+        context: &GuardContext,
+    ) -> Result<GuardResult> {
+        self.check_action(&GuardAction::NetworkEgress(host, port), context)
+            .await
+    }
+
+    /// Check a shell command action
+    pub async fn check_shell(&self, command: &str, context: &GuardContext) -> Result<GuardResult> {
+        self.check_action(&GuardAction::ShellCommand(command), context)
+            .await
+    }
+
+    /// Check an MCP tool invocation
+    pub async fn check_mcp_tool(
+        &self,
+        tool: &str,
+        args: &serde_json::Value,
+        context: &GuardContext,
+    ) -> Result<GuardResult> {
+        self.check_action(&GuardAction::McpTool(tool, args), context)
+            .await
+    }
+
+    /// Check untrusted text (e.g. fetched web content) for prompt-injection signals.
+    ///
+    /// This uses `GuardAction::Custom("untrusted_text", ...)` and is evaluated by `PromptInjectionGuard`.
+    pub async fn check_untrusted_text(
+        &self,
+        source: Option<&str>,
+        text: &str,
+        context: &GuardContext,
+    ) -> Result<GuardResult> {
+        let payload = match source {
+            Some(source) => serde_json::json!({ "source": source, "text": text }),
+            None => serde_json::json!({ "text": text }),
+        };
+
+        self.check_action(&GuardAction::Custom("untrusted_text", &payload), context)
+            .await
+    }
+
+    /// Check a patch action
+    pub async fn check_patch(
+        &self,
+        path: &str,
+        diff: &str,
+        context: &GuardContext,
+    ) -> Result<GuardResult> {
+        self.check_action(&GuardAction::Patch(path, diff), context)
+            .await
+    }
+
+    /// Check any action against all applicable guards
+    pub async fn check_action(
+        &self,
+        action: &GuardAction<'_>,
+        context: &GuardContext,
+    ) -> Result<GuardResult> {
+        Ok(self.check_action_report(action, context).await?.overall)
+    }
+
+    /// Check any action and return per-guard evidence plus the aggregated verdict.
+    pub async fn check_action_report(
+        &self,
+        action: &GuardAction<'_>,
+        context: &GuardContext,
+    ) -> Result<GuardReport> {
+        if let Some(msg) = self.config_error.as_ref() {
+            return Err(Error::ConfigError(msg.clone()));
+        }
+        if let Some(msg) = self.async_guard_init_error.as_ref() {
+            return Err(Error::ConfigError(msg.clone()));
+        }
+        let prepared = match self.prepare_origin_context(context, None).await? {
+            PreparedEvaluation::Continue(prepared) => *prepared,
+            PreparedEvaluation::Complete(report) => return Ok(*report),
+        };
+
+        self.check_action_report_prepared(action, prepared, None)
+            .await
+    }
+}
